+++ libNOM.io/PlatformMicrosoft.cs
@@ -19,9 +19,10 @@ public partial class PlatformMicrosoft : Platform
 
     internal static readonly string PATH = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), "Packages", "HelloGames.NoMansSky_bs190hzg1sesy", "SystemAppData", "wgs");
 
-    protected override int META_LENGTH_KNOWN => 0x14; // 20
+    protected override int META_LENGTH_KNOWN_VANILLA => 0x14; // 20
     internal override int META_LENGTH_TOTAL_VANILLA => 0x18; // 24
     internal override int META_LENGTH_TOTAL_WAYPOINT => 0x118; // 280
+    internal override int META_LENGTH_TOTAL_WORLDS => META_LENGTH_TOTAL_WAYPOINT; // no changes for this platform
 
     private const int BLOBCONTAINER_HEADER = 0x4; // 4
     private const int BLOBCONTAINER_COUNT = 0x2; // 2
@@ -29,12 +30,12 @@ public partial class PlatformMicrosoft : Platform
     private const int BLOBCONTAINER_TOTAL_LENGTH = sizeof(int) + sizeof(int) + BLOBCONTAINER_COUNT * (BLOBCONTAINER_IDENTIFIER_LENGTH + 2 * 0x10); // 328
 
     private const int CONTAINERSINDEX_HEADER = 0xE; // 14
-    private const long CONTAINERSINDEX_FOOTER = 0x10000000; // 268.435.456
+    private const long CONTAINERSINDEX_FOOTER = 0x10000000; // 268,435,456
     private const int CONTAINERSINDEX_OFFSET_BLOBCONTAINER_LIST = 0xC8; // 200
 
     internal static readonly byte[] SAVE_V2_HEADER = [.. Encoding.ASCII.GetBytes("HGSAVEV2"), 0x00];
     internal const int SAVE_V2_HEADER_PARTIAL_LENGTH = 0x8; // 8
-    internal const int SAVE_V2_CHUNK_MAX_LENGTH = 0x100000; // 1.048.576
+    internal const int SAVE_V2_CHUNK_MAX_LENGTH = 0x100000; // 1,048,576
 
     #endregion