+++ libNOM.io/PlatformPlaystation_FileSystemWatcher.cs
@@ -33,7 +33,7 @@ public partial class PlatformPlaystation : Platform
     /// </summary>
     private void RefreshContainerCollection()
     {
-        foreach (var metaIndex in Enumerable.Range(0, Constants.OFFSET_INDEX + COUNT_SAVES_TOTAL))
+        foreach (var metaIndex in Enumerable.Range(0, Constants.OFFSET_INDEX + MAX_SAVE_TOTAL))
             switch (metaIndex)
             {
                 case 0: